feat(SDK-4481) moves HPS constant to correct place
