+++ clevertap-hms/src/main/java/com/clevertap/android/hms/HmsConstants.java
@@ -1,14 +1,22 @@
 package com.clevertap.android.hms;
 
-interface HmsConstants {
+import com.clevertap.android.sdk.pushnotification.PushType;
 
-    String HMS_DELIVERY_TYPE = "hps";
-    String CT_HUAWEI_PROVIDER_CLASS = "com.clevertap.android.hms.HmsPushProvider";
-    String HUAWEI_SDK_CLASS = "com.huawei.hms.push.HmsMessageService";
-    String HPS_PROPERTY_REG_ID = "hps_token";
+public class HmsConstants {
 
-    String HMS_LOG_TAG = "HPS";
-    String HCM_SCOPE = "HCM";
-    String APP_ID_KEY = "client/app_id";
-    int MIN_CT_ANDROID_SDK_VERSION = 30800;
+    public static final String HMS_DELIVERY_TYPE = "hps";
+    public static final String HPS_PROPERTY_REG_ID = "hps_token";
+    public static final String CT_HUAWEI_PROVIDER_CLASS = "com.clevertap.android.hms.HmsPushProvider";
+    public static final String HUAWEI_SDK_CLASS = "com.huawei.hms.push.HmsMessageService";
+    public static final String HMS_LOG_TAG = "HPS";
+    public static final PushType HPS = new PushType(
+            HMS_DELIVERY_TYPE,
+            HPS_PROPERTY_REG_ID,
+            CT_HUAWEI_PROVIDER_CLASS,
+            HUAWEI_SDK_CLASS,
+            HMS_LOG_TAG
+    );
+    public static final String HCM_SCOPE = "HCM";
+    public static final String APP_ID_KEY = "client/app_id";
+    public static final int MIN_CT_ANDROID_SDK_VERSION = 30800;
 }
